Various changes

+++ src/main/java/com/miu/waafinalproject/model/responseDTO/PropertyResponseModel.java
@@ -16,7 +16,7 @@ public class PropertyResponseModel {
     Double price;
     String propertyType;
     Address address;
-    Short propertyViewCount;
+    long propertyViewCount;
     String image;
     Integer builtYear;
     String propertyStatus;
